refactor(bridge): extract message send
